requirements.txt
mcp[cli]
requests
openai
python-decouple
langchain
langchain-docling
langchain-openai
tiktoken
ipykernel
docling
tqdm
docling 
transformers
langchain-community
unstructured
pdfplumber
pi-heif
unstructured[local-inference]
chromadb
langchain_chroma
smolagents
LiteLLM
smolagents[mcp]